14.18.0 firebase auth state change, MAJOR-taken from shaun because of some error
