+++ src/App.js
@@ -1,0 +1,30 @@
+import { BrowserRouter, Route, Switch } from 'react-router-dom'
+
+// pages & components
+import Home from './pages/home/Home'
+import Login from './pages/login/Login'
+import Signup from './pages/signup/Signup'
+import Navbar from './components/Navbar'
+
+function App() {
+  return (
+    <div className="App">
+      <BrowserRouter>
+        <Navbar />
+        <Switch>
+          <Route exact path="/">
+            <Home />
+          </Route>
+          <Route path="/login">
+            <Login />
+          </Route>
+          <Route path="/signup">
+            <Signup />
+          </Route>
+        </Switch>
+      </BrowserRouter>
+    </div>
+  );
+}
+
+export default App
