--- src/App.jsx
@@ -1,32 +1,0 @@
-import "./App.css"
-import {
-  BrowserRouter as Router,
-  Routes,
-  Route,
-  NavLink,
-} from "react-router-dom"
-import Home from "./pages/home/Home"
-import Login from "./pages/login/Login"
-import Signup from "./pages/signup/Signup"
-import Navbar from "./component/navbar/navbar"
-
-function App() {
-  return (
-    <>
-      <Router>
-        <Navbar />
-        {/* <NavLink to="/">Home</NavLink>
-        <NavLink to="/login">Login</NavLink>
-        <NavLink to="/signup">Signup</NavLink> */}
-
-        <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/login" element={<Login />} />
-          <Route path="/signup" element={<Signup />} />
-        </Routes>
-      </Router>
-    </>
-  )
-}
-
-export default App
