+++ src/components/Navbar.module.css
@@ -17,8 +17,7 @@
   letter-spacing: 1px;
   font-size: 1.2em;
 }
-.navbar button,
-.navbar a {
+.navbar button, .navbar a {
   margin-left: 16px;
 }
 .navbar a {
