+++ src/hooks/useLogin.js
@@ -1,6 +1,6 @@
-import { useState, useEffect } from "react"
-import { projectAuth } from "../firebase/config"
-import { useAuthContext } from "./useAuthContext"
+import { useState, useEffect } from 'react'
+import { projectAuth } from '../firebase/config'
+import { useAuthContext } from './useAuthContext'
 
 export const useLogin = () => {
   const [isCancelled, setIsCancelled] = useState(false)
@@ -11,19 +11,20 @@ export const useLogin = () => {
   const login = async (email, password) => {
     setError(null)
     setIsPending(true)
-
+  
     try {
       // login
       const res = await projectAuth.signInWithEmailAndPassword(email, password)
 
       // dispatch login action
-      dispatch({ type: "LOGIN", payload: res.user })
+      dispatch({ type: 'LOGIN', payload: res.user })
 
       if (!isCancelled) {
         setIsPending(false)
         setError(null)
       }
-    } catch (err) {
+    } 
+    catch(err) {
       if (!isCancelled) {
         setError(err.message)
         setIsPending(false)
