+++ src/hooks/useLogout.js
@@ -1,13 +1,13 @@
-import { useEffect, useState } from "react"
-import { projectAuth } from "../firebase/config"
-import { useAuthContext } from "./useAuthContext"
+import { useEffect, useState } from 'react'
+import { projectAuth } from '../firebase/config'
+import { useAuthContext } from './useAuthContext'
 
 export const useLogout = () => {
   const [isCancelled, setIsCancelled] = useState(false)
   const [error, setError] = useState(null)
   const [isPending, setIsPending] = useState(false)
   const { dispatch } = useAuthContext()
-
+  
   const logout = async () => {
     setError(null)
     setIsPending(true)
@@ -15,16 +15,17 @@ export const useLogout = () => {
     try {
       // sign the user out
       await projectAuth.signOut()
-
+      
       // dispatch logout action
-      dispatch({ type: "LOGOUT" })
+      dispatch({ type: 'LOGOUT' })
 
       // update state
       if (!isCancelled) {
         setIsPending(false)
         setError(null)
-      }
-    } catch (err) {
+      } 
+    } 
+    catch(err) {
       if (!isCancelled) {
         setError(err.message)
         setIsPending(false)
