+++ src/hooks/useSignup.js
@@ -1,6 +1,6 @@
-import { useState, useEffect } from "react"
-import { projectAuth } from "../firebase/config"
-import { useAuthContext } from "./useAuthContext"
+import { useState, useEffect } from 'react'
+import { projectAuth } from '../firebase/config'
+import { useAuthContext } from './useAuthContext'
 
 export const useSignup = () => {
   const [isCancelled, setIsCancelled] = useState(false)
@@ -11,29 +11,27 @@ export const useSignup = () => {
   const signup = async (email, password, displayName) => {
     setError(null)
     setIsPending(true)
-
+  
     try {
       // signup
-      const res = await projectAuth.createUserWithEmailAndPassword(
-        email,
-        password
-      )
+      const res = await projectAuth.createUserWithEmailAndPassword(email, password)
 
       if (!res) {
-        throw new Error("Could not complete signup")
+        throw new Error('Could not complete signup')
       }
 
       // add display name to user
       await res.user.updateProfile({ displayName })
 
       // dispatch login action
-      dispatch({ type: "LOGIN", payload: res.user })
+      dispatch({ type: 'LOGIN', payload: res.user })
 
       if (!isCancelled) {
         setIsPending(false)
         setError(null)
       }
-    } catch (err) {
+    } 
+    catch(err) {
       if (!isCancelled) {
         setError(err.message)
         setIsPending(false)
