+++ src/index.css
@@ -1,20 +1,12 @@
-@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600&display=swap");
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;300;400;500;600;700;800;900&display=swap');
 
+/* base styles */
 body {
-  font-family: "Poppins", sans-serif;
+  font-family: Poppins, sans-serif;
   margin: 0;
-  padding: 0;
-  box-sizing: border-box;
   font-size: 1.1em;
 }
-
-ul,
-li,
-p,
-h1,
-h2,
-h3,
-h4 {
+ul, li, p, h1, h2, h3, h4 {
   margin: 0;
   padding: 0;
 }
